requirements.txt
Flask==3.0.0
groq==0.11.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
httpx==0.27.2
anyio==4.6.0
certifi==2024.8.30
